# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Product Service API","description":"CRUD API for products, built in Laboratory 2 of Distributed Systems. Storage is in memory and is lost when the service restarts.","contact":{"name":"Distributed Systems teaching team","url":"https://moodle.ensinolusofona.pt/"},"license":{"name":"Course material"},"version":"v1"},"servers":[{"url":"http://localhost:8081","description":"Local"}],"tags":[{"name":"Products","description":"Create, read, update and delete products"},{"name":"Suppliers","description":"Create, read, update and delete suppliers"}],"paths":{"/api/v1/suppliers/{id}":{"get":{"tags":["Suppliers"],"summary":"Get a supplier by id","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Supplier found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SupplierResponse"}}}},"400":{"description":"Malformed id","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Supplier not found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}}}},"put":{"tags":["Suppliers"],"summary":"Replace a supplier","description":"Full replace. The taxId must not belong to another supplier.","operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SupplierRequest"}}},"required":true},"responses":{"200":{"description":"Supplier updated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SupplierResponse"}}}},"400":{"description":"Validation failed","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Supplier not found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"409":{"description":"taxId already in use","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}}}},"delete":{"tags":["Suppliers"],"summary":"Delete a supplier","operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"Supplier deleted"},"400":{"description":"Malformed id","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Supplier not found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}}}}},"/api/v1/products/{id}":{"get":{"tags":["Products"],"summary":"Get a product by id","operationId":"getById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Product found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}},"404":{"description":"Product not found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}}}},"put":{"tags":["Products"],"summary":"Replace a product","description":"Full replace. The SKU must not belong to another product.","operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"Product updated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}},"400":{"description":"Validation failed","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Product not found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"409":{"description":"SKU already in use","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}}}},"delete":{"tags":["Products"],"summary":"Delete a product","operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"Product deleted"},"404":{"description":"Product not found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}}}}},"/api/v1/suppliers":{"get":{"tags":["Suppliers"],"summary":"List all suppliers","operationId":"getAll","responses":{"200":{"description":"Suppliers returned","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SupplierResponse"}}}}}}},"post":{"tags":["Suppliers"],"summary":"Create a supplier","description":"Creates a supplier and returns its location in the Location header. The taxId must not already be in use.","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SupplierRequest"}}},"required":true},"responses":{"201":{"description":"Supplier created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SupplierResponse"}}}},"400":{"description":"Validation failed","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"409":{"description":"taxId already in use","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}}}}},"/api/v1/products":{"get":{"tags":["Products"],"summary":"List all products","description":"Returns every product currently stored.","operationId":"getAll_1","responses":{"200":{"description":"Products returned","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"post":{"tags":["Products"],"summary":"Create a product","description":"Creates a product and returns its location in the Location header. The SKU must not already be in use.","operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"201":{"description":"Product created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}},"400":{"description":"Validation failed","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"409":{"description":"SKU already in use","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}}}}}},"components":{"schemas":{"SupplierRequest":{"type":"object","properties":{"companyName":{"type":"string","description":"Trading name","example":"Ibérica Componentes, Lda","maxLength":150,"minLength":2},"email":{"type":"string","format":"email","description":"Contact email","example":"encomendas@ibericacomponentes.pt","minLength":1},"taxId":{"type":"string","description":"Portuguese NIF, unique across all suppliers","example":"501234567","minLength":1,"pattern":"^[0-9]{9}$"},"phone":{"type":"string","description":"Contact phone","example":"+351 210 000 000","pattern":"^\\+?[0-9 ]{9,15}$"},"address":{"type":"string","description":"Street address","example":"Rua da Prata 120, 2.º Esq.","maxLength":200,"minLength":0},"city":{"type":"string","description":"City","example":"Lisboa","maxLength":100,"minLength":0},"postalCode":{"type":"string","description":"Portuguese postal code","example":"1100-052","minLength":1,"pattern":"^[0-9]{4}-[0-9]{3}$"},"country":{"type":"string","description":"ISO 3166-1 alpha-2 country code, defaults to PT","example":"PT","pattern":"^[A-Z]{2}$"},"active":{"type":"boolean","description":"Whether the supplier is active, defaults to true","example":true}},"required":["address","city","companyName","email","postalCode","taxId"]},"ProblemDetail":{"type":"object","properties":{"type":{"type":"string","format":"uri"},"title":{"type":"string"},"status":{"type":"integer","format":"int32"},"detail":{"type":"string"},"instance":{"type":"string","format":"uri"},"properties":{"type":"object","additionalProperties":{}}}},"SupplierResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"companyName":{"type":"string"},"email":{"type":"string"},"taxId":{"type":"string"},"phone":{"type":"string"},"address":{"type":"string"},"city":{"type":"string"},"postalCode":{"type":"string"},"country":{"type":"string"},"active":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"version":{"type":"integer","format":"int32"}}},"ProductRequest":{"type":"object","properties":{"name":{"type":"string","description":"Display name","example":"Mechanical Keyboard","maxLength":120,"minLength":3},"description":{"type":"string","description":"Longer description","example":"87-key, brown switches","maxLength":2000,"minLength":0},"sku":{"type":"string","description":"Stock keeping unit, unique across all products","example":"KB-87-BRN","minLength":1,"pattern":"^[A-Z0-9-]{3,40}$"},"price":{"type":"number","description":"Unit price","example":89.9,"minimum":0.00},"currency":{"type":"string","description":"ISO currency code, only EUR is accepted","example":"EUR","pattern":"^EUR$"},"stock":{"type":"integer","format":"int32","description":"Units in stock","example":25,"minimum":0}},"required":["name","price","sku","stock"]},"ProductResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"description":{"type":"string"},"sku":{"type":"string"},"price":{"type":"number"},"currency":{"type":"string"},"stock":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"version":{"type":"integer","format":"int32"}}}}}}